MGMT-11137 GPU driver pinning (#71)
* Add DriverVersion field in GPUAddon CRD

This field is optional and follows the conventions of the GPU operator,
in order to setup the driver container image to be used. So this field
is either an image tag (i.e. the semantic version of the driver, e.g.
515.48.07) or an image digest (i.e. `sha256:<digest>`).

- https://gitlab.com/nvidia/kubernetes/gpu-operator/-/blob/master/api/v1/clusterpolicy_types.go#L1340

Signed-off-by: Michail Resvanis <mresvani@redhat.com>

* Pin GPU driver version according to GPUAddon.Spec.DriverVersion

When the GPUAddon.Spec.DriverVersion is empty, then the driver-image from the
related images of the GPU operator CSV is defined as the
ClusterPolicy.Spec.Driver.Image.

When the GPUAddon.Spec.DriverVersion is non-empty, then this field is used
for the ClusterPolicy.Spec.Driver.Version.

Signed-off-by: Michail Resvanis <mresvani@redhat.com>